@@ -12,14 +12,14 @@
     "rogue-boi-core": "^0.1.0"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^2.5.2",
-    "svelte": "^4.0.0",
-    "vite": "^4.5.0",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
+    "svelte": "^5.0.0",
+    "vite": "^5.4.4",
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.10.0",
-    "eslint-plugin-svelte": "^2.35.1",
+    "eslint-plugin-svelte": "^2.45.1",
     "prettier": "^3.2.5",
-    "prettier-plugin-svelte": "^2.10.1"
+    "prettier-plugin-svelte": "^3.2.6"
   },
   "type": "module"
 }
