@@ -1,11 +1,10 @@
 <script>
   import { onDestroy, onMount } from "svelte";
   import { canvasStore } from "@rogueBoi/store.js";
-  import { writable } from "svelte/store";
 
-  let box;
-  let canvas;
-  let sizeStore = writable(720);
+  let box = $state();
+  let canvas = $state();
+  let size = $state(720);
   /** @type {ResizeObserver} */ let observer;
   onMount(() => {
     canvasStore.set(canvas);
@@ -13,7 +12,7 @@
       const w = box.offsetWidth;
       const h = box.offsetHeight;
 
-      sizeStore.set(Math.min(w, h));
+      size = Math.min(w, h);
     });
     observer.observe(box);
   });
@@ -21,20 +20,11 @@
     canvasStore.set(null);
     if (observer) observer.disconnect();
   });
-
-  let size;
-  $: {
-    size = $sizeStore;
-  }
 </script>
 
 <div class="parent" bind:this={box}>
-  <canvas
-    width={size}
-    height={size}
-    style="--width: {size}; --height: {size};"
-    bind:this={canvas}
-  />
+  <canvas width={size} height={size} style="--width: {size}; --height: {size};" bind:this={canvas}
+  ></canvas>
 </div>
 
 <style>
