@@ -1,11 +1,12 @@
 import App from "./App.svelte";
 import wasm, { initCore } from "rogue-boi-core";
+import { mount } from "svelte";
 
 const init = async () => {
   await wasm();
   const core = initCore();
 
-  new App({
+  mount(App, {
     target: document.body,
     props: {
       core
